Add files via upload
This is the current update of pinn_adaptive_mesh.ipynb. For some reason on my Jupyter Lab it was opening as pinn_adaptive_mesh (5).ipynb and was not updating automatically on Github. So I renamed the file with a _ at the end, created a new branch, and uploaded it here